+++ Dto/CollectionDtoResolverTrait.php
@@ -5,17 +5,17 @@ declare(strict_types=1);
 /*
  * This file is part of the DtoResolver package.
  *
- * (c) Wakeapp <https://wakeapp.ru>
+ * (c) MarfaTech <https://marfa-tech.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Wakeapp\Component\DtoResolver\Dto;
+namespace MarfaTech\Component\DtoResolver\Dto;
 
+use MarfaTech\Component\DtoResolver\Exception\FieldForIndexNotFoundException;
+use MarfaTech\Component\DtoResolver\Exception\InvalidCollectionItemException;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Wakeapp\Component\DtoResolver\Exception\FieldForIndexNotFoundException;
-use Wakeapp\Component\DtoResolver\Exception\InvalidCollectionItemException;
 
 use function key;
 use function next;