+++ Tests/Dto/DtoResolverTraitTest.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Wakeapp\Component\DtoResolver\Tests\Dto;
+namespace MarfaTech\Component\DtoResolver\Tests\Dto;
 
+use MarfaTech\Component\DtoResolver\Tests\Mock\DtoClass;
+use MarfaTech\Component\DtoResolver\Tests\Mock\TypedPropertyDtoClass;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\OptionsResolver\Exception\ExceptionInterface;
-use Wakeapp\Component\DtoResolver\Tests\Mock\DtoClass;
-use Wakeapp\Component\DtoResolver\Tests\Mock\TypedPropertyDtoClass;
 
 class DtoResolverTraitTest extends TestCase
 {