+++ Tests/Mock/TypedPropertyDtoClass.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Wakeapp\Component\DtoResolver\Tests\Mock;
+namespace MarfaTech\Component\DtoResolver\Tests\Mock;
 
-use Wakeapp\Component\DtoResolver\Dto\DtoResolverInterface;
-use Wakeapp\Component\DtoResolver\Dto\DtoResolverTrait;
+use MarfaTech\Component\DtoResolver\Dto\DtoResolverInterface;
+use MarfaTech\Component\DtoResolver\Dto\DtoResolverTrait;
 
 class TypedPropertyDtoClass implements DtoResolverInterface
 {